Compute input and output bindings

Computes the input- and output bindings for use in a causal map. Several heuristics may be used to determine the activities that are activated or consumed by an event. The Flexible Heuristic Miner (FHM) paper describes a heuristic that looks ahead (or looks back) until the end of the trace and determines those activities as activated for which no other cause (activity in a causal dependency) is found. This approach is implemented as type nearest.

causal_bindings(eventlog, dependencies, type = c("nearest"))

Arguments

eventlog
The bupaR event log.

dependencies
A dependency matrix obtained, for example, through dependency_matrix.

type
The heuristic used to determine the bindings. Currently only nearest is available.

Value

A data frame

Examples

causal_bindings(L_heur_1,
                dependencies = dependency_matrix(L_heur_1))

#>     next_act   act aid    case          start_time            end_time
#> 1          a Start 135 Case1.0 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 2          e     a 135 Case1.0 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 3        End     e 136 Case1.0 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 4       <NA>   End 136 Case1.0 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 5          a Start 130 Case1.1 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 6          e     a 130 Case1.1 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 7        End     e 131 Case1.1 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 8       <NA>   End 131 Case1.1 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 9          a Start 105 Case1.2 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 10         e     a 105 Case1.2 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 11       End     e 106 Case1.2 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 12      <NA>   End 106 Case1.2 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 13         a Start  96 Case1.3 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 14         e     a  96 Case1.3 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 15       End     e  97 Case1.3 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 16      <NA>   End  97 Case1.3 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 17         a Start 115 Case1.4 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 18         e     a 115 Case1.4 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 19       End     e 116 Case1.4 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 20      <NA>   End 116 Case1.4 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 21         a Start   1 Case2.0 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 22         b     a   1 Case2.0 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 23         c     b   2 Case2.0 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 24         e     c   3 Case2.0 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 25       End     e   4 Case2.0 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 26      <NA>   End   4 Case2.0 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 27         a Start  42 Case2.1 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 28         b     a  42 Case2.1 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 29         c     b  43 Case2.1 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 30         e     c  44 Case2.1 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 31       End     e  45 Case2.1 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 32      <NA>   End  45 Case2.1 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 33         a Start  46 Case2.2 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 34         b     a  46 Case2.2 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 35         c     b  47 Case2.2 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 36         e     c  48 Case2.2 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 37       End     e  49 Case2.2 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 38      <NA>   End  49 Case2.2 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 39         a Start  34 Case2.3 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 40         b     a  34 Case2.3 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 41         c     b  35 Case2.3 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 42         e     c  36 Case2.3 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 43       End     e  37 Case2.3 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 44      <NA>   End  37 Case2.3 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 45         a Start  38 Case2.4 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 46         b     a  38 Case2.4 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 47         c     b  39 Case2.4 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 48         e     c  40 Case2.4 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 49       End     e  41 Case2.4 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 50      <NA>   End  41 Case2.4 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 51         a Start  26 Case2.5 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 52         b     a  26 Case2.5 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 53         c     b  27 Case2.5 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 54         e     c  28 Case2.5 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 55       End     e  29 Case2.5 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 56      <NA>   End  29 Case2.5 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 57         a Start  30 Case2.6 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 58         b     a  30 Case2.6 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 59         c     b  31 Case2.6 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 60         e     c  32 Case2.6 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 61       End     e  33 Case2.6 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 62      <NA>   End  33 Case2.6 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 63         a Start  18 Case2.7 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 64         b     a  18 Case2.7 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 65         c     b  19 Case2.7 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 66         e     c  20 Case2.7 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 67       End     e  21 Case2.7 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 68      <NA>   End  21 Case2.7 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 69         a Start  22 Case2.8 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 70         b     a  22 Case2.8 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 71         c     b  23 Case2.8 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 72         e     c  24 Case2.8 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 73       End     e  25 Case2.8 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 74      <NA>   End  25 Case2.8 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 75         a Start  50 Case2.9 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 76         b     a  50 Case2.9 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 77         c     b  51 Case2.9 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 78         e     c  52 Case2.9 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 79       End     e  53 Case2.9 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 80      <NA>   End  53 Case2.9 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 81         a Start  84 Case3.0 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 82         c     a  84 Case3.0 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 83         b     c  85 Case3.0 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 84         e     b  86 Case3.0 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 85       End     e  87 Case3.0 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 86      <NA>   End  87 Case3.0 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 87         a Start  80 Case3.1 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 88         c     a  80 Case3.1 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 89         b     c  81 Case3.1 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 90         e     b  82 Case3.1 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 91       End     e  83 Case3.1 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 92      <NA>   End  83 Case3.1 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 93         a Start  92 Case3.2 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 94         c     a  92 Case3.2 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 95         b     c  93 Case3.2 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 96         e     b  94 Case3.2 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 97       End     e  95 Case3.2 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 98      <NA>   End  95 Case3.2 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 99         a Start  88 Case3.3 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 100        c     a  88 Case3.3 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 101        b     c  89 Case3.3 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 102        e     b  90 Case3.3 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 103      End     e  91 Case3.3 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 104     <NA>   End  91 Case3.3 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 105        a Start 107 Case3.4 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 106        c     a 107 Case3.4 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 107        b     c 108 Case3.4 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 108        e     b 109 Case3.4 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 109      End     e 110 Case3.4 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 110     <NA>   End 110 Case3.4 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 111        a Start  98 Case3.5 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 112        c     a  98 Case3.5 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 113        b     c  99 Case3.5 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 114        e     b 100 Case3.5 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 115      End     e 101 Case3.5 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 116     <NA>   End 101 Case3.5 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 117        a Start 117 Case3.6 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 118        c     a 117 Case3.6 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 119        b     c 118 Case3.6 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 120        e     b 119 Case3.6 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 121      End     e 120 Case3.6 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 122     <NA>   End 120 Case3.6 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 123        a Start 111 Case3.7 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 124        c     a 111 Case3.7 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 125        b     c 112 Case3.7 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 126        e     b 113 Case3.7 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 127      End     e 114 Case3.7 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 128     <NA>   End 114 Case3.7 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 129        a Start  69 Case3.8 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 130        c     a  69 Case3.8 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 131        b     c  70 Case3.8 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 132        e     b  71 Case3.8 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 133      End     e  72 Case3.8 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 134     <NA>   End  72 Case3.8 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 135        a Start  73 Case3.9 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 136        c     a  73 Case3.9 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 137        b     c  74 Case3.9 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 138        e     b  75 Case3.9 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 139      End     e  76 Case3.9 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 140     <NA>   End  76 Case3.9 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 141        a Start 102 Case4.0 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 142        b     a 102 Case4.0 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 143        e     b 103 Case4.0 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 144      End     e 104 Case4.0 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 145     <NA>   End 104 Case4.0 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 146        a Start  77 Case5.0 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 147        c     a  77 Case5.0 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 148        e     c  78 Case5.0 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 149      End     e  79 Case5.0 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 150     <NA>   End  79 Case5.0 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 151        a Start 137 Case6.0 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 152        d     a 137 Case6.0 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 153        e     d 138 Case6.0 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 154      End     e 139 Case6.0 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 155     <NA>   End 139 Case6.0 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 156        a Start 132 Case6.1 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 157        d     a 132 Case6.1 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 158        e     d 133 Case6.1 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 159      End     e 134 Case6.1 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 160     <NA>   End 134 Case6.1 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 161        a Start 127 Case6.2 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 162        d     a 127 Case6.2 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 163        e     d 128 Case6.2 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 164      End     e 129 Case6.2 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 165     <NA>   End 129 Case6.2 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 166        a Start 124 Case6.3 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 167        d     a 124 Case6.3 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 168        e     d 125 Case6.3 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 169      End     e 126 Case6.3 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 170     <NA>   End 126 Case6.3 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 171        a Start 121 Case6.4 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 172        d     a 121 Case6.4 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 173        e     d 122 Case6.4 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 174      End     e 123 Case6.4 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 175     <NA>   End 123 Case6.4 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 176        a Start  63 Case6.5 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 177        d     a  63 Case6.5 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 178        e     d  64 Case6.5 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 179      End     e  65 Case6.5 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 180     <NA>   End  65 Case6.5 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 181        a Start  66 Case6.6 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 182        d     a  66 Case6.6 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 183        e     d  67 Case6.6 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 184      End     e  68 Case6.6 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 185     <NA>   End  68 Case6.6 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 186        a Start  57 Case6.7 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 187        d     a  57 Case6.7 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 188        e     d  58 Case6.7 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 189      End     e  59 Case6.7 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 190     <NA>   End  59 Case6.7 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 191        a Start  60 Case6.8 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 192        d     a  60 Case6.8 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 193        e     d  61 Case6.8 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 194      End     e  62 Case6.8 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 195     <NA>   End  62 Case6.8 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 196        a Start  54 Case6.9 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 197        d     a  54 Case6.9 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 198        e     d  55 Case6.9 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 199      End     e  56 Case6.9 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 200     <NA>   End  56 Case6.9 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 201        a Start  10 Case7.0 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 202        d     a  10 Case7.0 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 203        d     d  11 Case7.0 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 204        e     d  12 Case7.0 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 205      End     e  13 Case7.0 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 206     <NA>   End  13 Case7.0 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 207        a Start  14 Case7.1 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 208        d     a  14 Case7.1 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 209        d     d  15 Case7.1 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 210        e     d  16 Case7.1 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 211      End     e  17 Case7.1 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 212     <NA>   End  17 Case7.1 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 213        a Start   5 Case8.0 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 214        d     a   5 Case8.0 2010-11-04 19:57:15 2010-11-04 19:57:15
#> 215        d     d   6 Case8.0 2010-11-04 19:58:15 2010-11-04 19:58:15
#> 216        d     d   7 Case8.0 2010-11-04 19:59:15 2010-11-04 19:59:15
#> 217        e     d   8 Case8.0 2010-11-04 20:00:15 2010-11-04 20:00:15
#> 218      End     e   9 Case8.0 2010-11-04 20:01:15 2010-11-04 20:01:15
#> 219     <NA>   End   9 Case8.0 2010-11-04 20:01:15 2010-11-04 20:01:15
#>     min_order     next_start_time       next_end_time from_id to_id
#> 1        -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 2         135 2010-11-04 19:58:15 2010-11-04 19:58:15       1     5
#> 3         136 2010-11-04 19:58:15 2010-11-04 19:58:15       5     6
#> 4         Inf                <NA>                <NA>       6    NA
#> 5        -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 6         130 2010-11-04 19:58:15 2010-11-04 19:58:15       1     5
#> 7         131 2010-11-04 19:58:15 2010-11-04 19:58:15       5     6
#> 8         Inf                <NA>                <NA>       6    NA
#> 9        -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 10        105 2010-11-04 19:58:15 2010-11-04 19:58:15       1     5
#> 11        106 2010-11-04 19:58:15 2010-11-04 19:58:15       5     6
#> 12        Inf                <NA>                <NA>       6    NA
#> 13       -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 14         96 2010-11-04 19:58:15 2010-11-04 19:58:15       1     5
#> 15         97 2010-11-04 19:58:15 2010-11-04 19:58:15       5     6
#> 16        Inf                <NA>                <NA>       6    NA
#> 17       -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 18        115 2010-11-04 19:58:15 2010-11-04 19:58:15       1     5
#> 19        116 2010-11-04 19:58:15 2010-11-04 19:58:15       5     6
#> 20        Inf                <NA>                <NA>       6    NA
#> 21       -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 22          1 2010-11-04 19:58:15 2010-11-04 19:58:15       1     2
#> 23          2 2010-11-04 19:59:15 2010-11-04 19:59:15       2     3
#> 24          3 2010-11-04 20:00:15 2010-11-04 20:00:15       3     5
#> 25          4 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 26        Inf                <NA>                <NA>       6    NA
#> 27       -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 28         42 2010-11-04 19:58:15 2010-11-04 19:58:15       1     2
#> 29         43 2010-11-04 19:59:15 2010-11-04 19:59:15       2     3
#> 30         44 2010-11-04 20:00:15 2010-11-04 20:00:15       3     5
#> 31         45 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 32        Inf                <NA>                <NA>       6    NA
#> 33       -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 34         46 2010-11-04 19:58:15 2010-11-04 19:58:15       1     2
#> 35         47 2010-11-04 19:59:15 2010-11-04 19:59:15       2     3
#> 36         48 2010-11-04 20:00:15 2010-11-04 20:00:15       3     5
#> 37         49 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 38        Inf                <NA>                <NA>       6    NA
#> 39       -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 40         34 2010-11-04 19:58:15 2010-11-04 19:58:15       1     2
#> 41         35 2010-11-04 19:59:15 2010-11-04 19:59:15       2     3
#> 42         36 2010-11-04 20:00:15 2010-11-04 20:00:15       3     5
#> 43         37 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 44        Inf                <NA>                <NA>       6    NA
#> 45       -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 46         38 2010-11-04 19:58:15 2010-11-04 19:58:15       1     2
#> 47         39 2010-11-04 19:59:15 2010-11-04 19:59:15       2     3
#> 48         40 2010-11-04 20:00:15 2010-11-04 20:00:15       3     5
#> 49         41 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 50        Inf                <NA>                <NA>       6    NA
#> 51       -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 52         26 2010-11-04 19:58:15 2010-11-04 19:58:15       1     2
#> 53         27 2010-11-04 19:59:15 2010-11-04 19:59:15       2     3
#> 54         28 2010-11-04 20:00:15 2010-11-04 20:00:15       3     5
#> 55         29 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 56        Inf                <NA>                <NA>       6    NA
#> 57       -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 58         30 2010-11-04 19:58:15 2010-11-04 19:58:15       1     2
#> 59         31 2010-11-04 19:59:15 2010-11-04 19:59:15       2     3
#> 60         32 2010-11-04 20:00:15 2010-11-04 20:00:15       3     5
#> 61         33 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 62        Inf                <NA>                <NA>       6    NA
#> 63       -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 64         18 2010-11-04 19:58:15 2010-11-04 19:58:15       1     2
#> 65         19 2010-11-04 19:59:15 2010-11-04 19:59:15       2     3
#> 66         20 2010-11-04 20:00:15 2010-11-04 20:00:15       3     5
#> 67         21 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 68        Inf                <NA>                <NA>       6    NA
#> 69       -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 70         22 2010-11-04 19:58:15 2010-11-04 19:58:15       1     2
#> 71         23 2010-11-04 19:59:15 2010-11-04 19:59:15       2     3
#> 72         24 2010-11-04 20:00:15 2010-11-04 20:00:15       3     5
#> 73         25 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 74        Inf                <NA>                <NA>       6    NA
#> 75       -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 76         50 2010-11-04 19:58:15 2010-11-04 19:58:15       1     2
#> 77         51 2010-11-04 19:59:15 2010-11-04 19:59:15       2     3
#> 78         52 2010-11-04 20:00:15 2010-11-04 20:00:15       3     5
#> 79         53 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 80        Inf                <NA>                <NA>       6    NA
#> 81       -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 82         84 2010-11-04 19:58:15 2010-11-04 19:58:15       1     3
#> 83         85 2010-11-04 19:59:15 2010-11-04 19:59:15       3     2
#> 84         86 2010-11-04 20:00:15 2010-11-04 20:00:15       2     5
#> 85         87 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 86        Inf                <NA>                <NA>       6    NA
#> 87       -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 88         80 2010-11-04 19:58:15 2010-11-04 19:58:15       1     3
#> 89         81 2010-11-04 19:59:15 2010-11-04 19:59:15       3     2
#> 90         82 2010-11-04 20:00:15 2010-11-04 20:00:15       2     5
#> 91         83 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 92        Inf                <NA>                <NA>       6    NA
#> 93       -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 94         92 2010-11-04 19:58:15 2010-11-04 19:58:15       1     3
#> 95         93 2010-11-04 19:59:15 2010-11-04 19:59:15       3     2
#> 96         94 2010-11-04 20:00:15 2010-11-04 20:00:15       2     5
#> 97         95 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 98        Inf                <NA>                <NA>       6    NA
#> 99       -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 100        88 2010-11-04 19:58:15 2010-11-04 19:58:15       1     3
#> 101        89 2010-11-04 19:59:15 2010-11-04 19:59:15       3     2
#> 102        90 2010-11-04 20:00:15 2010-11-04 20:00:15       2     5
#> 103        91 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 104       Inf                <NA>                <NA>       6    NA
#> 105      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 106       107 2010-11-04 19:58:15 2010-11-04 19:58:15       1     3
#> 107       108 2010-11-04 19:59:15 2010-11-04 19:59:15       3     2
#> 108       109 2010-11-04 20:00:15 2010-11-04 20:00:15       2     5
#> 109       110 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 110       Inf                <NA>                <NA>       6    NA
#> 111      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 112        98 2010-11-04 19:58:15 2010-11-04 19:58:15       1     3
#> 113        99 2010-11-04 19:59:15 2010-11-04 19:59:15       3     2
#> 114       100 2010-11-04 20:00:15 2010-11-04 20:00:15       2     5
#> 115       101 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 116       Inf                <NA>                <NA>       6    NA
#> 117      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 118       117 2010-11-04 19:58:15 2010-11-04 19:58:15       1     3
#> 119       118 2010-11-04 19:59:15 2010-11-04 19:59:15       3     2
#> 120       119 2010-11-04 20:00:15 2010-11-04 20:00:15       2     5
#> 121       120 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 122       Inf                <NA>                <NA>       6    NA
#> 123      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 124       111 2010-11-04 19:58:15 2010-11-04 19:58:15       1     3
#> 125       112 2010-11-04 19:59:15 2010-11-04 19:59:15       3     2
#> 126       113 2010-11-04 20:00:15 2010-11-04 20:00:15       2     5
#> 127       114 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 128       Inf                <NA>                <NA>       6    NA
#> 129      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 130        69 2010-11-04 19:58:15 2010-11-04 19:58:15       1     3
#> 131        70 2010-11-04 19:59:15 2010-11-04 19:59:15       3     2
#> 132        71 2010-11-04 20:00:15 2010-11-04 20:00:15       2     5
#> 133        72 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 134       Inf                <NA>                <NA>       6    NA
#> 135      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 136        73 2010-11-04 19:58:15 2010-11-04 19:58:15       1     3
#> 137        74 2010-11-04 19:59:15 2010-11-04 19:59:15       3     2
#> 138        75 2010-11-04 20:00:15 2010-11-04 20:00:15       2     5
#> 139        76 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 140       Inf                <NA>                <NA>       6    NA
#> 141      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 142       102 2010-11-04 19:58:15 2010-11-04 19:58:15       1     2
#> 143       103 2010-11-04 19:59:15 2010-11-04 19:59:15       2     5
#> 144       104 2010-11-04 19:59:15 2010-11-04 19:59:15       5     6
#> 145       Inf                <NA>                <NA>       6    NA
#> 146      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 147        77 2010-11-04 19:58:15 2010-11-04 19:58:15       1     3
#> 148        78 2010-11-04 19:59:15 2010-11-04 19:59:15       3     5
#> 149        79 2010-11-04 19:59:15 2010-11-04 19:59:15       5     6
#> 150       Inf                <NA>                <NA>       6    NA
#> 151      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 152       137 2010-11-04 19:58:15 2010-11-04 19:58:15       1     4
#> 153       138 2010-11-04 19:59:15 2010-11-04 19:59:15       4     5
#> 154       139 2010-11-04 19:59:15 2010-11-04 19:59:15       5     6
#> 155       Inf                <NA>                <NA>       6    NA
#> 156      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 157       132 2010-11-04 19:58:15 2010-11-04 19:58:15       1     4
#> 158       133 2010-11-04 19:59:15 2010-11-04 19:59:15       4     5
#> 159       134 2010-11-04 19:59:15 2010-11-04 19:59:15       5     6
#> 160       Inf                <NA>                <NA>       6    NA
#> 161      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 162       127 2010-11-04 19:58:15 2010-11-04 19:58:15       1     4
#> 163       128 2010-11-04 19:59:15 2010-11-04 19:59:15       4     5
#> 164       129 2010-11-04 19:59:15 2010-11-04 19:59:15       5     6
#> 165       Inf                <NA>                <NA>       6    NA
#> 166      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 167       124 2010-11-04 19:58:15 2010-11-04 19:58:15       1     4
#> 168       125 2010-11-04 19:59:15 2010-11-04 19:59:15       4     5
#> 169       126 2010-11-04 19:59:15 2010-11-04 19:59:15       5     6
#> 170       Inf                <NA>                <NA>       6    NA
#> 171      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 172       121 2010-11-04 19:58:15 2010-11-04 19:58:15       1     4
#> 173       122 2010-11-04 19:59:15 2010-11-04 19:59:15       4     5
#> 174       123 2010-11-04 19:59:15 2010-11-04 19:59:15       5     6
#> 175       Inf                <NA>                <NA>       6    NA
#> 176      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 177        63 2010-11-04 19:58:15 2010-11-04 19:58:15       1     4
#> 178        64 2010-11-04 19:59:15 2010-11-04 19:59:15       4     5
#> 179        65 2010-11-04 19:59:15 2010-11-04 19:59:15       5     6
#> 180       Inf                <NA>                <NA>       6    NA
#> 181      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 182        66 2010-11-04 19:58:15 2010-11-04 19:58:15       1     4
#> 183        67 2010-11-04 19:59:15 2010-11-04 19:59:15       4     5
#> 184        68 2010-11-04 19:59:15 2010-11-04 19:59:15       5     6
#> 185       Inf                <NA>                <NA>       6    NA
#> 186      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 187        57 2010-11-04 19:58:15 2010-11-04 19:58:15       1     4
#> 188        58 2010-11-04 19:59:15 2010-11-04 19:59:15       4     5
#> 189        59 2010-11-04 19:59:15 2010-11-04 19:59:15       5     6
#> 190       Inf                <NA>                <NA>       6    NA
#> 191      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 192        60 2010-11-04 19:58:15 2010-11-04 19:58:15       1     4
#> 193        61 2010-11-04 19:59:15 2010-11-04 19:59:15       4     5
#> 194        62 2010-11-04 19:59:15 2010-11-04 19:59:15       5     6
#> 195       Inf                <NA>                <NA>       6    NA
#> 196      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 197        54 2010-11-04 19:58:15 2010-11-04 19:58:15       1     4
#> 198        55 2010-11-04 19:59:15 2010-11-04 19:59:15       4     5
#> 199        56 2010-11-04 19:59:15 2010-11-04 19:59:15       5     6
#> 200       Inf                <NA>                <NA>       6    NA
#> 201      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 202        10 2010-11-04 19:58:15 2010-11-04 19:58:15       1     4
#> 203        11 2010-11-04 19:59:15 2010-11-04 19:59:15       4     4
#> 204        12 2010-11-04 20:00:15 2010-11-04 20:00:15       4     5
#> 205        13 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 206       Inf                <NA>                <NA>       6    NA
#> 207      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 208        14 2010-11-04 19:58:15 2010-11-04 19:58:15       1     4
#> 209        15 2010-11-04 19:59:15 2010-11-04 19:59:15       4     4
#> 210        16 2010-11-04 20:00:15 2010-11-04 20:00:15       4     5
#> 211        17 2010-11-04 20:00:15 2010-11-04 20:00:15       5     6
#> 212       Inf                <NA>                <NA>       6    NA
#> 213      -Inf 2010-11-04 19:57:15 2010-11-04 19:57:15       7     1
#> 214         5 2010-11-04 19:58:15 2010-11-04 19:58:15       1     4
#> 215         6 2010-11-04 19:59:15 2010-11-04 19:59:15       4     4
#> 216         7 2010-11-04 20:00:15 2010-11-04 20:00:15       4     4
#> 217         8 2010-11-04 20:01:15 2010-11-04 20:01:15       4     5
#> 218         9 2010-11-04 20:01:15 2010-11-04 20:01:15       5     6
#> 219       Inf                <NA>                <NA>       6    NA
#>     possible_output possible_input binding_output binding_input
#> 1                 a                             a              
#> 2           b, c, d          Start                        Start
#> 3               End        b, c, d            End              
#> 4                                e                            e
#> 5                 a                             a              
#> 6           b, c, d          Start                        Start
#> 7               End        b, c, d            End              
#> 8                                e                            e
#> 9                 a                             a              
#> 10          b, c, d          Start                        Start
#> 11              End        b, c, d            End              
#> 12                               e                            e
#> 13                a                             a              
#> 14          b, c, d          Start                        Start
#> 15              End        b, c, d            End              
#> 16                               e                            e
#> 17                a                             a              
#> 18          b, c, d          Start                        Start
#> 19              End        b, c, d            End              
#> 20                               e                            e
#> 21                a                             a              
#> 22          b, c, d          Start           b, c         Start
#> 23                e              a              e             a
#> 24                e              a              e             a
#> 25              End        b, c, d            End          b, c
#> 26                               e                            e
#> 27                a                             a              
#> 28          b, c, d          Start           b, c         Start
#> 29                e              a              e             a
#> 30                e              a              e             a
#> 31              End        b, c, d            End          b, c
#> 32                               e                            e
#> 33                a                             a              
#> 34          b, c, d          Start           b, c         Start
#> 35                e              a              e             a
#> 36                e              a              e             a
#> 37              End        b, c, d            End          b, c
#> 38                               e                            e
#> 39                a                             a              
#> 40          b, c, d          Start           b, c         Start
#> 41                e              a              e             a
#> 42                e              a              e             a
#> 43              End        b, c, d            End          b, c
#> 44                               e                            e
#> 45                a                             a              
#> 46          b, c, d          Start           b, c         Start
#> 47                e              a              e             a
#> 48                e              a              e             a
#> 49              End        b, c, d            End          b, c
#> 50                               e                            e
#> 51                a                             a              
#> 52          b, c, d          Start           b, c         Start
#> 53                e              a              e             a
#> 54                e              a              e             a
#> 55              End        b, c, d            End          b, c
#> 56                               e                            e
#> 57                a                             a              
#> 58          b, c, d          Start           b, c         Start
#> 59                e              a              e             a
#> 60                e              a              e             a
#> 61              End        b, c, d            End          b, c
#> 62                               e                            e
#> 63                a                             a              
#> 64          b, c, d          Start           b, c         Start
#> 65                e              a              e             a
#> 66                e              a              e             a
#> 67              End        b, c, d            End          b, c
#> 68                               e                            e
#> 69                a                             a              
#> 70          b, c, d          Start           b, c         Start
#> 71                e              a              e             a
#> 72                e              a              e             a
#> 73              End        b, c, d            End          b, c
#> 74                               e                            e
#> 75                a                             a              
#> 76          b, c, d          Start           b, c         Start
#> 77                e              a              e             a
#> 78                e              a              e             a
#> 79              End        b, c, d            End          b, c
#> 80                               e                            e
#> 81                a                             a              
#> 82          b, c, d          Start           b, c         Start
#> 83                e              a              e             a
#> 84                e              a              e             a
#> 85              End        b, c, d            End          b, c
#> 86                               e                            e
#> 87                a                             a              
#> 88          b, c, d          Start           b, c         Start
#> 89                e              a              e             a
#> 90                e              a              e             a
#> 91              End        b, c, d            End          b, c
#> 92                               e                            e
#> 93                a                             a              
#> 94          b, c, d          Start           b, c         Start
#> 95                e              a              e             a
#> 96                e              a              e             a
#> 97              End        b, c, d            End          b, c
#> 98                               e                            e
#> 99                a                             a              
#> 100         b, c, d          Start           b, c         Start
#> 101               e              a              e             a
#> 102               e              a              e             a
#> 103             End        b, c, d            End          b, c
#> 104                              e                            e
#> 105               a                             a              
#> 106         b, c, d          Start           b, c         Start
#> 107               e              a              e             a
#> 108               e              a              e             a
#> 109             End        b, c, d            End          b, c
#> 110                              e                            e
#> 111               a                             a              
#> 112         b, c, d          Start           b, c         Start
#> 113               e              a              e             a
#> 114               e              a              e             a
#> 115             End        b, c, d            End          b, c
#> 116                              e                            e
#> 117               a                             a              
#> 118         b, c, d          Start           b, c         Start
#> 119               e              a              e             a
#> 120               e              a              e             a
#> 121             End        b, c, d            End          b, c
#> 122                              e                            e
#> 123               a                             a              
#> 124         b, c, d          Start           b, c         Start
#> 125               e              a              e             a
#> 126               e              a              e             a
#> 127             End        b, c, d            End          b, c
#> 128                              e                            e
#> 129               a                             a              
#> 130         b, c, d          Start           b, c         Start
#> 131               e              a              e             a
#> 132               e              a              e             a
#> 133             End        b, c, d            End          b, c
#> 134                              e                            e
#> 135               a                             a              
#> 136         b, c, d          Start           b, c         Start
#> 137               e              a              e             a
#> 138               e              a              e             a
#> 139             End        b, c, d            End          b, c
#> 140                              e                            e
#> 141               a                             a              
#> 142         b, c, d          Start              b         Start
#> 143               e              a              e             a
#> 144             End        b, c, d            End             b
#> 145                              e                            e
#> 146               a                             a              
#> 147         b, c, d          Start              c         Start
#> 148               e              a              e             a
#> 149             End        b, c, d            End             c
#> 150                              e                            e
#> 151               a                             a              
#> 152         b, c, d          Start              d         Start
#> 153               e              a              e             a
#> 154             End        b, c, d            End             d
#> 155                              e                            e
#> 156               a                             a              
#> 157         b, c, d          Start              d         Start
#> 158               e              a              e             a
#> 159             End        b, c, d            End             d
#> 160                              e                            e
#> 161               a                             a              
#> 162         b, c, d          Start              d         Start
#> 163               e              a              e             a
#> 164             End        b, c, d            End             d
#> 165                              e                            e
#> 166               a                             a              
#> 167         b, c, d          Start              d         Start
#> 168               e              a              e             a
#> 169             End        b, c, d            End             d
#> 170                              e                            e
#> 171               a                             a              
#> 172         b, c, d          Start              d         Start
#> 173               e              a              e             a
#> 174             End        b, c, d            End             d
#> 175                              e                            e
#> 176               a                             a              
#> 177         b, c, d          Start              d         Start
#> 178               e              a              e             a
#> 179             End        b, c, d            End             d
#> 180                              e                            e
#> 181               a                             a              
#> 182         b, c, d          Start              d         Start
#> 183               e              a              e             a
#> 184             End        b, c, d            End             d
#> 185                              e                            e
#> 186               a                             a              
#> 187         b, c, d          Start              d         Start
#> 188               e              a              e             a
#> 189             End        b, c, d            End             d
#> 190                              e                            e
#> 191               a                             a              
#> 192         b, c, d          Start              d         Start
#> 193               e              a              e             a
#> 194             End        b, c, d            End             d
#> 195                              e                            e
#> 196               a                             a              
#> 197         b, c, d          Start              d         Start
#> 198               e              a              e             a
#> 199             End        b, c, d            End             d
#> 200                              e                            e
#> 201               a                             a              
#> 202         b, c, d          Start              d         Start
#> 203               e              a              e             a
#> 204               e              a              e             a
#> 205             End        b, c, d            End             d
#> 206                              e                            e
#> 207               a                             a              
#> 208         b, c, d          Start              d         Start
#> 209               e              a              e             a
#> 210               e              a              e             a
#> 211             End        b, c, d            End             d
#> 212                              e                            e
#> 213               a                             a              
#> 214         b, c, d          Start              d         Start
#> 215               e              a              e             a
#> 216               e              a              e             a
#> 217               e              a              e             a
#> 218             End        b, c, d            End             d
#> 219                              e                            e