Frequency map profile

Function to create a frequency profile for a process map.

causal_frequency(
  value = c("absolute", "relative"),
  color_scale = "PuBu",
  color_edges = "dodgerblue4"
)

Arguments

value
The type of frequency value to be used: absolute, relative (percentage of activity instances).

color_scale
Name of color scale to be used for nodes. Defaults to PuBu. See Rcolorbrewer::brewer.pal.info() for all options.

color_edges
The color used for edges. Defaults to dodgerblue4.

Examples

causal_net(L_heur_1,
           type = causal_frequency("relative"))

#> Nodes
#> # A tibble: 7 x 12
#>   act   from_id     n n_distinct_cases bindings_input bindings_output label     
#>   <chr>   <int> <dbl>            <dbl> <list>         <list>          <chr>     
#> 1 a           1    40               40 <list [1]>     <list [4]>      "a\n28.78~
#> 2 b           2    21               21 <list [1]>     <list [1]>      "b\n15.11~
#> 3 c           3    21               21 <list [1]>     <list [1]>      "c\n15.11~
#> 4 d           4    17               13 <list [1]>     <list [1]>      "d\n12.23~
#> 5 e           5    40               40 <list [4]>     <list [1]>      "e\n28.78~
#> 6 End         6    40               40 <list [1]>     <list [0]>      "End"     
#> 7 Start       7    40               40 <list [0]>     <list [1]>      "Start"   
#> # ... with 5 more variables: color_level <dbl>, shape <chr>, fontcolor <chr>,
#> #   color <chr>, tooltip <chr>
#> Edges
#> # A tibble: 8 x 8
#>   antecedent consequent   dep from_id to_id     n label  penwidth
#>   <chr>      <chr>      <dbl>   <int> <int> <dbl> <chr>     <dbl>
#> 1 Start      a          0.976       7     1    40 100%       5   
#> 2 a          b          0.917       1     2    21 38.18%     2.53
#> 3 a          c          0.917       1     3    21 38.18%     2.53
#> 4 a          d          0.929       1     4    13 23.64%     1.95
#> 5 b          e          0.917       2     5    21 100%       5   
#> 6 c          e          0.917       3     5    21 100%       5   
#> 7 d          e          0.929       4     5    17 100%       5   
#> 8 e          End        0.976       5     6    40 100%       5