Create a dependency matrix

Creates a dependency matrix from a precedence matrix (precedence_matrix) based on different approaches.

dependency_matrix(
  eventlog = NULL,
  dependency_type = dependency_type_fhm(threshold_dependency = threshold,
    threshold_frequency = threshold_frequency, ...),
  threshold = 0.9,
  threshold_frequency = 0,
  ...
)

Arguments

eventlog
A bupaR event log, may be NULL when a precedence matrix is provided.

dependency_type
Which approach to use for calculation of the dependency matrix. Currently only (dependency_type_fhm) is available.

threshold
A dependency threshold, usually in the interval [0,1], filtering out dependencies below the threshold.

threshold_frequency
An absolute frequency threshold filtering dependencies which are observed infrequently.

...
Parameters forwarded to (dependency_type_fhm).

Value

A square matrix with class dependency_matrix containing the computed dependency values between all activities.

See also

Examples

d <- dependency_matrix(L_heur_1)
print(d)

#>           consequent
#> antecedent         a         b         c         d         e       End Start
#>      a     0.0000000 0.9166667 0.9166667 0.9285714 0.0000000 0.0000000     0
#>      b     0.0000000 0.0000000 0.0000000 0.0000000 0.9166667 0.0000000     0
#>      c     0.0000000 0.0000000 0.0000000 0.0000000 0.9166667 0.0000000     0
#>      d     0.0000000 0.0000000 0.0000000 0.0000000 0.9285714 0.0000000     0
#>      e     0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.9756098     0
#>      End   0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>      Start 0.9756098 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#> attr(,"class")
#> [1] "dependency_matrix" "matrix"            "array"            

as.matrix(d)

#>           consequent
#> antecedent         a         b         c         d         e       End Start
#>      a     0.0000000 0.9166667 0.9166667 0.9285714 0.0000000 0.0000000     0
#>      b     0.0000000 0.0000000 0.0000000 0.0000000 0.9166667 0.0000000     0
#>      c     0.0000000 0.0000000 0.0000000 0.0000000 0.9166667 0.0000000     0
#>      d     0.0000000 0.0000000 0.0000000 0.0000000 0.9285714 0.0000000     0
#>      e     0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.9756098     0
#>      End   0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#>      Start 0.9756098 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000     0
#> attr(,"class")
#> [1] "dependency_matrix" "matrix"            "array"