Dependency type based on time intervals

Computes the dependencies based on the approach taking into account activity durations based on life-cycle transitions.

dependency_type_lifecycle(
  threshold_dependency = 0.9,
  threshold_l1 = threshold_dependency,
  threshold_frequency = 0,
  all_connected = FALSE,
  endpoints_connected = FALSE
)

Arguments

threshold_dependency

A dependency threshold, usually in the interval [0,1], filtering out dependencies below the threshold.

threshold_l1

A dependency threshold, usually in the interval [0,1], filtering out self-loop dependencies below the threshold.

threshold_frequency

An absolute frequency threshold filtering dependencies which are observed infrequently.

all_connected

If TRUE the best antecedent and consequent (as determined by the dependency measure) are going to be added regardless of the threshold value.

endpoints_connected

If TRUE the start/end activity is added as antecedent/consequent when an activity would not be connected according to the threshold value.

Value

A dependency type.

References

A. Burattin and A. Sperduti, “Heuristics Miner for Time Intervals,” in ESANN 2010, 18th European Symposium on Artificial Neural Networks, Bruges, Belgium, April 28-30, 2010, Proceedings, 2010.

Examples

dependency_matrix(L_heur_1,
                  dependency_type = dependency_type_fhm(all_connected = TRUE))
#>           consequent
#> antecedent       End Start         a         b         c         d         e
#>      End   0.0000000     0 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>      Start 0.0000000     0 0.9756098 0.0000000 0.0000000 0.0000000 0.0000000
#>      a     0.0000000     0 0.0000000 0.9166667 0.9166667 0.9285714 0.0000000
#>      b     0.0000000     0 0.0000000 0.0000000 0.0000000 0.0000000 0.9166667
#>      c     0.0000000     0 0.0000000 0.0000000 0.0000000 0.0000000 0.9166667
#>      d     0.0000000     0 0.0000000 0.0000000 0.0000000 0.0000000 0.9285714
#>      e     0.9756098     0 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "dependency_matrix" "matrix"            "array"