Custom map profile

Function to create a custom map profile based on some event log attribute.

causal_custom(FUN = mean, attribute, units = "",
  color_scale = "RdPu", color_edges = "red4", ...)

Arguments

FUN
A summary function to be called on the process time of a specific activity, e.g. mean, median, min, max

attribute
The name of the case attribute to visualize (should be numeric)

units
Character to be placed after values (e.g. EUR for monitary euro values)

color_scale
Name of color scale to be used for nodes. Defaults to RdPu See Rcolorbrewer::brewer.pal.info() for all options.

color_edges
The color used for edges. Defaults to red4.

...
Additional arguments forwarded to FUN

Details

If used for edges, it will show the attribute values which related to the out-going node of the edge.#'

Examples

causal_net(L_heur_1,
           type_nodes = causal_custom(attribute = "timestamp"),
           type_edges = causal_custom(attribute = "timestamp"))

#> Nodes
#> # A tibble: 7 x 10
#>   act   from_id bindings_input bindings_output label color_level shape fontcolor
#>   <chr>   <int> <list>         <list>          <chr>       <dbl> <chr> <chr>    
#> 1 a           1 <list [1]>     <list [4]>      "a\n… 1288900636. rect… white    
#> 2 b           2 <list [1]>     <list [1]>      "b\n… 1288900724. rect… white    
#> 3 c           3 <list [1]>     <list [1]>      "c\n… 1288900724. rect… white    
#> 4 d           4 <list [1]>     <list [1]>      "d\n… 1288900713. rect… white    
#> 5 e           5 <list [4]>     <list [1]>      "e\n… 1288900784. rect… white    
#> 6 End         6 <list [1]>     <list [0]>      "End"          0  circ… brown4   
#> 7 Start       7 <list [0]>     <list [1]>      "Sta…          0  circ… chartreu…
#> # … with 2 more variables: color <chr>, tooltip <chr>
#> Edges
#> # A tibble: 8 x 10
#>   antecedent consequent   dep from_id to_id     n  value label_numeric label
#>   <chr>      <chr>      <dbl>   <int> <int> <dbl>  <dbl>         <dbl> <chr>
#> 1 Start      a          0.976       7     1    40 0.                0  " "  
#> 2 a          b          0.917       1     2    21 1.29e9   1288900636. "128…
#> 3 a          c          0.917       1     3    21 1.29e9   1288900636. "128…
#> 4 a          d          0.929       1     4    13 1.29e9   1288900636. "128…
#> 5 b          e          0.917       2     5    21 1.29e9   1288900724. "128…
#> 6 c          e          0.917       3     5    21 1.29e9   1288900724. "128…
#> 7 d          e          0.929       4     5    17 1.29e9   1288900713. "128…
#> 8 e          End        0.975       5     6    40 1.29e9   1288900784. "128…
#> # … with 1 more variable: penwidth <dbl>

#> Warning: Prefixing `UQ()` with the rlang namespace is deprecated as of rlang 0.3.0.
#> Please use the non-prefixed form or `!!` instead.
#> 
#>   # Bad:
#>   rlang::expr(mean(rlang::UQ(var) * 100))
#> 
#>   # Ok:
#>   rlang::expr(mean(UQ(var) * 100))
#> 
#>   # Good:
#>   rlang::expr(mean(!!var * 100))
#> 
#> This warning is displayed once per session.