Dependency type based on Flexible Heuristics Miner (FHM)

Computes the dependencies based on the approach known as Flexible Heuristics Miner.

dependency_type_fhm(threshold_dependency = 0.9,
  threshold_l1 = threshold_dependency,
  threshold_l2 = threshold_dependency, threshold_frequency = 0,
  all_connected = FALSE)

Arguments

threshold_dependency
A dependency threshold, usually in the interval [0,1], filtering out dependencies below the threshold.

threshold_l1
A dependency threshold, usually in the interval [0,1], filtering out self-loop dependencies below the threshold.

threshold_l2
A dependency threshold, usually in the interval [0,1], filtering out length-two loop dependencies below the threshold.

threshold_frequency
An absolute frequency threshold filtering dependencies which are observed infrequently.

all_connected
If TRUE the best antecedent and consequent (as determined by the dependency measure) are going to be added regardless of the threshold value.

Value

A dependency type.

References

A. J. M. M. Weijters and J. T. S. Ribeiro, "Flexible Heuristics Miner (FHM)," 2011 IEEE Symposium on Computational Intelligence and Data Mining (CIDM), Paris, 2011, pp. 310-317. doi: 10.1109/CIDM.2011.5949453

Examples

dependency_matrix(L_heur_1,
                  dependency_type = dependency_type_fhm(all_connected = TRUE))

#>           consequent
#> antecedent         a         b         c         d         e   End Start
#>      a     0.0000000 0.9166667 0.9166667 0.9285714 0.0000000 0.000     0
#>      b     0.0000000 0.0000000 0.0000000 0.0000000 0.9166667 0.000     0
#>      c     0.0000000 0.0000000 0.0000000 0.0000000 0.9166667 0.000     0
#>      d     0.0000000 0.0000000 0.0000000 0.0000000 0.9285714 0.000     0
#>      e     0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.975     0
#>      End   0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.000     0
#>      Start 0.9756098 0.0000000 0.0000000 0.0000000 0.0000000 0.000     0
#> attr(,"class")
#> [1] "dependency_matrix" "matrix"